add and try use token str arena

+++ include/lreng.h
@@ -4,7 +4,7 @@
 
 dynarr_t tokenize(
     const char* src,
-    const int src_len,
+    const unsigned long src_len,
     const int is_debug
 );
